Add a cache_utilities module #682
Refactor-out the cache_df decorator.
